+++ opennlp-api/src/main/java/opennlp/tools/tokenize/SubwordPiece.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package opennlp.subword;
+package opennlp.tools.tokenize;
 
 import opennlp.tools.util.Span;
