+++ opennlp-extensions/opennlp-subword/src/test/java/opennlp/subword/sentencepiece/SentencePieceRealModelEvalTest.java
@@ -26,7 +26,7 @@ import java.util.stream.Stream;
 
 import org.junit.jupiter.api.Test;
 
-import opennlp.subword.SubwordPiece;
+import opennlp.tools.tokenize.SubwordPiece;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
